// src/types/Config.type.ts
export type ConfigDefaultType = {
    title?: string;
    show_today?: boolean;
    show_overdue?: boolean;
    show_future?: number;
    show_all?: boolean;

}

export type ConfigType = {
    title?: string;
    show_today: boolean;
    show_overdue: boolean;
    show_future: number;
    show_all?: boolean;

}